# openapi.yaml
openapi: 3.0.0
info:
  title: CRM & Inventory Management API
  description: |
    RESTful API for managing customers, products, orders, and inventory.
    
    ## Features
    - Customer management (CRM)
    - Product catalog management
    - Order processing and tracking
    - Inventory management
    
    ## Authentication
    Currently using mock data. Production version will require JWT authentication.
    
  version: 1.0.0
  contact:
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:5000
    description: Development server
  - url: http://localhost:5000/api
    description: API base path

tags:
  - name: Health
    description: Health check and system status
  - name: Customers
    description: Customer management operations
  - name: Products
    description: Product catalog operations
  - name: Orders
    description: Order management and processing

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  message:
                    type: string
                    example: API is running

  /api/customers:
    get:
      tags:
        - Customers
      summary: Get all customers
      description: Retrieve a paginated list of all customers
      operationId: getCustomers
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by customer status
          required: false
          schema:
            type: string
            enum: [active, inactive, pending]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerSummary'
                  total:
                    type: integer
                    example: 100
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 50
    
    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Add a new customer to the system
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 3
                  message:
                    type: string
                    example: Customer created successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/{customerId}:
    get:
      tags:
        - Customers
      summary: Get customer by ID
      description: Retrieve detailed information about a specific customer
      operationId: getCustomerById
      parameters:
        - name: customerId
          in: path
          description: Customer ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve a paginated list of all products
      operationId: getProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: category
          in: query
          description: Filter by product category
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductSummary'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    
    post:
      tags:
        - Products
      summary: Create a new product
      description: Add a new product to the catalog
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Invalid input

  /api/products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve detailed information about a specific product
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

  /api/orders:
    get:
      tags:
        - Orders
      summary: Get all orders
      description: Retrieve a paginated list of all orders
      operationId: getOrders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [pending, processing, completed, cancelled]
        - name: customer_id
          in: query
          description: Filter by customer ID
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderSummary'
                  total:
                    type: integer
                  page:
                    type: integer
    
    post:
      tags:
        - Orders
      summary: Create a new order
      description: Place a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
        '400':
          description: Invalid input

  /api/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve detailed information about a specific order
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

components:
  schemas:
    CustomerSummary:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        status:
          type: string
          enum: [active, inactive, pending]
          example: active

    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerSummary'
        - type: object
          properties:
            phone:
              type: string
              example: 555-0123
            company:
              type: string
              example: Acme Corp
            address:
              type: string
              example: 123 Main St, City, State 12345
            created_at:
              type: string
              format: date-time
              example: 2024-01-15T10:30:00Z

    CustomerCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: 555-0123
        company:
          type: string
          example: Acme Corp
        address:
          type: string
          example: 123 Main St

    ProductSummary:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Widget A
        price:
          type: number
          format: float
          example: 29.99
        stock:
          type: integer
          example: 100
        category:
          type: string
          example: Electronics

    Product:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          properties:
            description:
              type: string
              example: High-quality electronic widget
            sku:
              type: string
              example: WGT-A-001

    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
          example: Widget C
        price:
          type: number
          format: float
          example: 39.99
        stock:
          type: integer
          example: 75
        category:
          type: string
          example: Electronics
        description:
          type: string
          example: New electronic widget

    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer_id:
          type: integer
          example: 1
        total:
          type: number
          format: float
          example: 299.99
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
          example: completed
        order_date:
          type: string
          format: date
          example: 2024-02-01

    Order:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            shipping_address:
              type: string
              example: 123 Main St, City, State 12345
            payment_method:
              type: string
              example: Credit Card

    OrderItem:
      type: object
      properties:
        product_id:
          type: integer
          example: 1
        product_name:
          type: string
          example: Widget A
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 29.99
        subtotal:
          type: number
          format: float
          example: 59.98

    OrderCreate:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: integer
          example: 1
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
                example: 1
              quantity:
                type: integer
                example: 2
        shipping_address:
          type: string
          example: 123 Main St
        payment_method:
          type: string
          example: Credit Card

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid input
        message:
          type: string
          example: The request body is missing required fields